playing sound issue
I have tried to get sound to work. I found the chrome policy on autoplay. So I am trying to attach it to a button but it doesn't seem to be working. However, the object "audio" is being created. I can type audio.play() into the browser console and it plays the sound. Any ideas?