bug report implementation added
I allowed issues to be opened. So we can give feedback to one another based on pull requests. Go to the repo click issues. Click new issue. To attach a pull request or assign the issue to a specific person look in the right pane under the issues tab. It saves automatically. Then we can review there.